+++ src/mass_driver/models/status.py
@@ -2,12 +2,9 @@
 
 
 from enum import Enum
+from traceback import format_exception
 
 from pydantic import BaseModel
-
-from mass_driver.models.forge import PRResult
-from mass_driver.models.patchdriver import PatchResult
-from mass_driver.models.repository import ClonedRepo, SourcedRepo
 
 ScanResult = dict[str, dict]
 """The output of one or more scanner(s) on a single repo, indexed by scanner-name"""
@@ -23,20 +20,37 @@ class Phase(str, Enum):
     FORGE = "FORGE"
 
 
-class RepoOutcome(BaseModel):
-    """A single repo's processing status"""
+class ExceptionRecord(BaseModel):
+    """A serialized version of an Exception"""
 
-    repo_id: str
-    """The Repo's identifier among this run. Unique-ish string"""
-    status: Phase
-    """The status of the repo's processing"""
-    source: SourcedRepo
-    """The Repo's info from Source"""
-    clone: ClonedRepo | None = None
-    """The cloned repo's data if any"""
-    scan: ScanResult | None = None
-    """The result of a Scan, if any"""
-    patch: PatchResult | None = None
-    """The result of a Scan, if any"""
-    forge: PRResult | None = None
-    """The result of Forge, if any"""
+    exception_type: str
+    """The type of exception thrown"""
+    exception_details: str
+    """The string version of the exception"""
+    backtrace: list[str]
+    """The exception backtrace as array of string"""
+
+    @classmethod
+    def from_exception(cls, e: Exception):
+        """Create an ExceptionRecord object from Exception object"""
+        return cls(
+            exception_type=e.__class__.__name__,
+            exception_details=str(e),
+            backtrace=format_exception(e),
+        )
+
+
+class Error(BaseModel):
+    """Captures an issue during a specific phase"""
+
+    activity: Phase
+    """During which activity did the error occur"""
+    # context: str
+    # """A short string explaining what was attempted, that failed"""
+    error: ExceptionRecord
+    """The record of the exception, what went wrong"""
+
+    @classmethod
+    def from_exception(cls, activity: Phase, exception: Exception):
+        """Create an Error from Exception object"""
+        return cls(activity=activity, error=ExceptionRecord.from_exception(exception))