+++ src/mass_driver/tests/test_followup.py
@@ -6,7 +6,6 @@ import pytest
 from mass_driver.forge_run import PROutcome
 from mass_driver.forges.dummy import DUMMY_PR_URL
 from mass_driver.models.patchdriver import PatchOutcome
-from mass_driver.models.status import Phase
 from mass_driver.tests.fixtures import (
     copy_folder,
     massdrive_runlocal,
@@ -14,7 +13,6 @@ from mass_driver.tests.fixtures import (
 )
 
 
-@pytest.mark.xfail(reason="2-mig,1forge known bug")
 def test_2migration_but_1forge(tmp_path, shared_datadir, monkeypatch):
     """Scenario: Forge a single time when 2 migration but only one fails"""
     # Given two repos to migrate
@@ -32,30 +30,29 @@ def test_2migration_but_1forge(tmp_path, shared_datadir, monkeypatch):
 
     activityconfig_filepath = shared_datadir / "2mig_1forge.toml"
     # When I run mass-driver with 2 migrations, 1 forge
-    result = massdrive_runlocal(None, activityconfig_filepath)
-    assert result is not None, "Should have a result"
-    assert result.migration_result is not None, "Should have a migration result"
-    mig1 = result.migration_result[repo1_id]
-    mig2 = result.migration_result[repo2_id]
+    results = massdrive_runlocal(None, activityconfig_filepath)
+    assert results is not None, "Should have a result"
+    result1 = results.repos[repo1_id]
+    assert result1.patch is not None, "Should have a migration result"
+    result2 = results.repos[repo2_id]
+    mig1 = result1.patch
     # Then I get OK outcome on migration1
     assert mig1.outcome == PatchOutcome.PATCHED_OK, "Wrong outcome from patching"
     # And I get ALREADY_PATCHED outcome on migration2
+    mig2 = result2.patch
+    assert mig2 is not None, "Should have a migration result"
     assert mig2.outcome == PatchOutcome.ALREADY_PATCHED, "Wrong outcome from patching"
 
-    if result.forge_result is None:
-        return pytest.fail("Should have a forge result")
+    assert result1.forge is not None, "Should have a forge result"
     # And I get only 1 forge result
-    assert len(result.forge_result.keys()) == 1, "Should get only one forge result"
-    forge1 = result.forge_result[repo1_id]
-    if forge1 is None:
-        return pytest.fail("Should have a forge result for mig1")
-
+    forge1 = result1.forge
+    assert forge1 is not None, "Should have a forge result for mig1"
     assert forge1.outcome == PROutcome.PR_CREATED, "Should succeed creating PR"
     # And I get the PR URL I want
     assert forge1.pr_html_url == DUMMY_PR_URL, "Should have returned correct PR URL"
 
 
-@pytest.mark.xfail(reason="clonefail, sent for PR known bug")
+@pytest.mark.xfail(reason="Clone1 failure, bad test for now? known bug too")
 def test_1badclone_but_1forge(tmp_path, shared_datadir, monkeypatch):
     """Scenario: Forge a single time given 2 migration but 1 failed cloning"""
     # Given two repos to migrate
@@ -69,25 +66,24 @@ def test_1badclone_but_1forge(tmp_path, shared_datadir, monkeypatch):
 
     activityconfig_filepath = shared_datadir / "2mig_1forge.toml"
     # When I run mass-driver with 2 migrations to forge
-    result = massdrive_runlocal(None, activityconfig_filepath)
-    assert result is not None, "Should have a result"
-    assert result.migration_result is not None, "Should have a migration result"
-    mig1 = result.migration_result[repo1_id]
-    mig2 = result.migration_result[repo2_id]
+    results = massdrive_runlocal(None, activityconfig_filepath)
+    assert results is not None, "Should have a result"
+    result1 = results.repos[repo1_id]
+    assert result1.error is None, "Should have no error for mig1"
+    assert result1.patch is not None, "Should have a migration result"
+    result2 = results.repos[repo2_id]
+    mig1 = result1.patch
     # Then I get OK outcome on migration1
     assert mig1.outcome == PatchOutcome.PATCHED_OK, "Wrong outcome from patching"
-    # And I get Clone error outcome on migration2
+    # And I get ALREADY_PATCHED outcome on migration2
+    mig2 = result2.patch
+    assert mig2 is not None, "Should have a migration result"
     assert mig2.outcome == PatchOutcome.PATCH_ERROR, "Wrong outcome from patching"
-    error = mig2.error
-    assert error is not None, "Should have gotten a record of the error"
-    assert error.activity == Phase.CLONE, "Error should be marked from Clone activity"
 
+    assert result1.forge is not None, "Should have a forge result"
     # And I get only 1 forge result
-    assert result.forge_result is not None, "Should have a forge result"
-    assert len(result.forge_result.keys()) == 1, "Should get only one forge result"
-    forge1 = result.forge_result[repo1_id]
+    forge1 = result1.forge
     assert forge1 is not None, "Should have a forge result for mig1"
-
     assert forge1.outcome == PROutcome.PR_CREATED, "Should succeed creating PR"
     # And I get the PR URL I want
     assert forge1.pr_html_url == DUMMY_PR_URL, "Should have returned correct PR URL"